+++ cmd/plugins.go
@@ -7,4 +7,5 @@ import (
 	_ "github.com/rapatel0/alpha/internal/ext/askuser"
 	_ "github.com/rapatel0/alpha/internal/ext/todo"
 	_ "github.com/rapatel0/alpha/internal/ext/tokenspeed"
+	_ "github.com/rapatel0/alpha/internal/ext/toolstats"
 )